Compress segmentation data modules to one file
