--- src/netspresso_trainer/dataloaders/segmentation/__init__.py
@@ -1,3 +1,0 @@
-from .dataset import SegmentationSampleLoader
-from .huggingface import SegmentationHFDataset
-from .local import SegmentationCustomDataset
